refactor: extract Ctx lifecycle methods, complete context.go split
